requirements.txt
google-genai
edge-tts
requests
moviepy
python-dotenv
pydub
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
tiktok-uploader
opencv-python
